% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sigma_gamma.R
\name{calculate_sigma_gamma}
\alias{calculate_sigma_gamma}
\title{Calculate sigma and gamma using distributions IT, SRT, IBST}
\usage{
calculate_sigma_gamma(IT, SRT, IBST)
}
\arguments{
\item{IT}{array-like}

\item{SRT}{array-like}

\item{IBST}{array-like}
}
\value{
a list:
\describe{
  \item{sigma}{float}
  \item{gamma}{float}
}
}
\description{
Calculate sigma and gamma using distributions IT, SRT, IBST
}
