% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const_trim.R
\name{const_trim}
\alias{const_trim}
\title{Trims leading values in numerical vector}
\usage{
const_trim(v, threshold)
}
\arguments{
\item{v}{a numerical vector}

\item{threshold}{the threshold used for cutting the vector}
}
\value{
the first index of the vector which entry grows more than \code{threshold}, starting from 1 to \code{length(v)}
}
\description{
Removes the first elements of a numerical vector till they do show
an increase higher than threshold.
}
\examples{
\dontrun{
	const_trim(c(0,0,0,1,1.3,1.7,2,3,4.5,6,9,12),1)
}
}
