% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_start_end.R
\name{detect_start_end}
\alias{detect_start_end}
\title{Detects start and end of outbreak
 
... more specifics ...}
\usage{
detect_start_end(v, threshold = 0.005)
}
\arguments{
\item{threshold}{}
}
\value{
a list of start and end indexes
}
\description{
Detects start and end of outbreak
 
... more specifics ...
}
\examples{
\dontrun{

}
 
}
