% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R0.R
\name{estimate_R0}
\alias{estimate_R0}
\title{Estimate R0}
\usage{
estimate_R0(sigma, gamma, I, time_start, time_end)
}
\arguments{
\item{sigma}{float}

\item{gamma}{float}

\item{I}{array-like}

\item{time_start}{int time serie of which to calculate R0}

\item{time_end}{int}
}
\value{
a list containing: R0, start, end. start and end are the extremes of J (see details). Ignore if any is passed by user.
}
\description{
Estimate R0 of a time series I using an exponential approximation of I (I_approx) in early stages of diffusion.
The method used is a least square fitting of I_approx over I.
}
\details{
R0 will be estimated using I[J], where J is:
	(time_start:time_end) 	if user provides both 
	(1:time_end) 			if user provides time_end but not time_start
	(time_start:length(I))  if user provides time_start but not time_end
	(start:end)             if user provides neither. start and end are extrapolated with exp_period
								and aim to identify the extremes of the stage of exponential growth in I
}
