% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe_fit.R
\name{exe_fit}
\alias{exe_fit}
\title{Fit of the data and forecast via the parameters estimed}
\usage{
exe_fit(sample_cases, sample_date, days)
}
\arguments{
\item{sample_cases}{values for growth fitting.}

\item{sample_date}{sample dates as serial consecutive dates (should be included in \code{days}).}

\item{days}{full vector of serial dates.}
}
\value{
a list of fitting results
\describe{
  \item{out_fit}{The output of the fitting performed by \link[growthcurver]{SummarizeGrowth}.}
  \item{fittedPoints}{A data frame containing dates and corresponding fitted values listed by columns}
  \item{out_chisq}{The output of the chi-squared independence test of residuals}
}
}
\description{
Fit of the data and forecast via the parameters estimed
}
\examples{
\dontrun{
  data <- get_regionTS()
  data <- subset(data, denominazione_regione == "Lombardia")
  exe_fit(data$totale_casi, data$data_seriale, c(1:50))
}

}
