% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe_fit_TS.R
\name{exe_fit_TS}
\alias{exe_fit_TS}
\title{Times curve fitting and forecasting}
\usage{
exe_fit_TS(sample_cases_TS, p, I, q, TS_t)
}
\arguments{
\item{sample_cases_TS}{values for ARIMA FITTING.}

\item{p}{autoregressive parameter.}

\item{I}{integration order}

\item{q}{MA order}

\item{TS_t}{full lenght of the forecasting.}
}
\value{
a list of fitting results. Specifically:
\describe{
  \item{out_fit_TS}{The output of the the fit the ARIMA model (p,I,q)}
}
}
\description{
Times curve fitting and forecasting
}
\examples{
\dontrun{
  data <- get_regionTS()
  data <- subset(data, denominazione_regione == "Lombardia")
  exe_fit_TS(data$totale_casi, data$data_seriale, c(1:50))
}

}
