% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generates densities}
\usage{
generate(distr, parameters)
}
\description{
Generates densities given distribution names and parameters.
Generates distribution without values of 3th quartile (trade off: accuracy - data)
}
