% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agecases.R
\name{get_agecases}
\alias{get_agecases}
\title{Data acquisition of cases age region from Epicentro ISS}
\usage{
get_agecases(regNames)
}
\arguments{
\item{regNames}{vector. Name of italian regions}
}
\value{
list. Return a list of region containing a list of dataframe
}
\description{
Data acquisition of cases age from Epicentro ISS \href{https://www.epicentro.iss.it/}{(Link)} of summary information and age cases per region
}
\examples{
\dontrun{
get_agecases(regNames)
}

}
