% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countryTS.R
\name{get_countryTS}
\alias{get_countryTS}
\title{Data acquisition of Italy}
\usage{
get_countryTS()
}
\value{
dataframe. Return a dataframe containing the dataset covid19 for Italty
}
\description{
Data acquisition of Italy
}
\examples{
\dontrun{
get_countryTS()
}

}
