% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provTS.R
\name{get_provTS}
\alias{get_provTS}
\title{Data acquisition sorted by province}
\usage{
get_provTS()
}
\value{
A list containing the original dataset splitted into single provinces.
}
\description{
Data acquisition sorted by province
}
\examples{
\dontrun{
  names(get_provTS()) # List of provinces names
}

}
