% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regionTS.R
\name{get_regionTS}
\alias{get_regionTS}
\title{Data acquisition sorted by region}
\usage{
get_regionTS()
}
\value{
A list containing the original dataset splitted into each region.
}
\description{
Data acquisition sorted by region
}
\examples{
\dontrun{
regTS <- get_regionTS()
names(regTS) # List of regions names
}

}
