% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperatures.R
\name{get_temperatures}
\alias{get_temperatures}
\title{Italy temperatures' getter}
\usage{
get_temperatures()
}
\value{
A data frame with three columns: \code{loc} (locality), \code{date}, \code{temp} (temperature).
}
\description{
Downloads mean temperatures in Italy from \url{https://www.ilmeteo.it}, throughout the period
2020/02/24 - today and put them in a table after rearranging and interpolating missing values.
}
\details{
A search of the mean temperatures of the 107 italian provinces is performed. The available data is stored
and manipulated as follows: missing dates are added; missing temperature values are linearly interpolated 
through the preceding and following values of the same province; a weighted mean of the mean temperatures
of each day, with provinces' areas as weights, is performed; its result is interpreted as the mean temperature
of the whole country and is appended to the data frame.
}
\examples{
\dontrun{
 temperatures <- get_temperatures()
 subset(temperatures, loc == "Italy")
}
}
