% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensivecare_cap_data.R
\docType{data}
\name{intensivecare_cap}
\alias{intensivecare_cap}
\title{Dataset intensive care capacity}
\format{
tibble
}
\usage{
intensivecare_cap
}
\description{
Dataset of intensive care capacity per Region in Italy
}
\keyword{datasets}
