% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_factotum.R
\name{SEIR_factotum}
\alias{SEIR_factotum}
\title{SEIR factotum V2}
\usage{
SEIR_factotum(
  P,
  R,
  N,
  end,
  normalise = TRUE,
  time_step = 1,
  future = 0,
  distr_IT = "exp",
  distr_SRT = "exp",
  distr_IBST = "none",
  par_IT = list(rate = 1/5.2),
  par_SRT = list(rate = log(2)/5),
  par_IBST = list(),
  R0_exp_est_start = 1,
  R0_exp_est_end = 5,
  R0_msp,
  R0_stages = c(),
  fit_Rt = FALSE,
  plot_data = TRUE
)
}
\description{
Solves a normalised (population=1) simplified SEIR system (no immunity loss, no mortality):
}
\details{
1_ extract the S, E, I, R, information from the input data P (positive) and R using the distributions IT, SRT, IBST. 
The default distribution for IT is an exponential with rate 1/5.2 (5.2 being the mean Incubation Time)
	https://www.nejm.org/doi/full/10.1056/NEJMoa2001316
The default distribution for SRT is an exponential with rate log(2)/5 (5 being the median time from symptomaticity to removal - hospitalisation)
	https://www.epicentro.iss.it/coronavirus/bollettino/Report-COVID-2019_13_aprile.pdf
The default distribution for IBST is an none as no data is available
	
2_ parameters sigma and gamma are estimated as 1/mean(IT) and 1/mean(SRT) resp

3_a if the user inputs a valid Rt (consisting of R0_msp and R0_stagesof length RN and RN-1 resp), SEIR_factotum will solve the multi-stage SEIR with 
parameters (at n-th stage) sigma, gamma, beta[n]=Rt[n]*gamma. It will return U and sol, U being the refined data and sol being 
the solution of the SEIR equation. 

3_b if no valid Rt is passed, SEIR_factotum will estimate an initial R0 using asymptotic properties of the series I in the early stages
(from R0_exp_est_start to R0_exp_est_end, meaning R0 exponential estimate start/end), 
then it will solve a single stage SEIR system with parameters sigma, gamma, beta=R0*gamma. 
This can be used to show how the growth would have been like if no measures had taken place.
It will return U and sol as above, plus result of the estimation of R0
}
