% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGUI.R
\name{sGUI}
\alias{sGUI}
\title{Generate GUI interface}
\usage{
sGUI()
}
\value{
shiny app
}
\description{
GUI function

This app allows to generate a GUI that plots the data of COVID19
}
\examples{
\dontrun{
sGUI()
}
}
