% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe_fit.R
\name{exe_fit}
\alias{exe_fit}
\title{Logistic curve fitting}
\usage{
exe_fit(sample_cases, sample_date, days)
}
\arguments{
\item{sample_cases}{values for growth fitting.}

\item{sample_date}{sample dates as serial consecutive dates (should be included in \code{days}).}

\item{days}{full vector of serial dates.}
}
\value{
a list of fitting results. Specifically:
\describe{
  \item{out_fit}{The output of the fitting performed by \link[growthcurver]{SummarizeGrowth}.}
  \item{fittedPoints}{A data frame containing dates and corresponding fitted values of cumulative cases, listed by columns.}
  \item{out_chisq}{The output of the chi-squared independence test of residuals.}
  \item{fittedPoints_der}{A data frame containing dates and the corresponding fitted values of new cases, listed by columns.}
}
}
\description{
Logistic curve fitting
}
\examples{
\dontrun{
  data <- get_regionTS()
  data <- subset(data, denominazione_regione == "Lombardia")
  exe_fit(data$totale_casi, data$data_seriale, c(1:50))
}

}
