% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intesivecare_cap.R
\name{get_intensivecare_cap}
\alias{get_intensivecare_cap}
\title{Data acquisition regional insentive care capacity}
\usage{
get_intensivecare_cap(regionTS)
}
\arguments{
\item{regionTS}{list. List of the Italy regional dataset}
}
\value{
dataframe. Return a dataframe containing information about regional intensive care capacity
}
\description{
Data acquisition regional insentive care capacity
}
\examples{
\dontrun{
regionTS = get_regionTS()
get_intensivecare_cap(regionTS)
}

}
