% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runC19.R
\name{runC19}
\alias{runC19}
\title{Shiny app with exploratory analysis of covid19}
\usage{
runC19()
}
\value{
Runs the Shiny app retreiving the data, running the analysis and showing the report with interactive visualisations.
}
\description{
The function works with no arguments. The function depends on the dataset
}
